Separate ld signature namespace
