+++ src/main/java/com/apicatalog/vc/integrity/DataIntegritySuite.java
@@ -6,9 +6,9 @@ import com.apicatalog.ld.DocumentError;
 import com.apicatalog.ld.schema.LdSchema;
 import com.apicatalog.ld.schema.LdTerm;
 import com.apicatalog.ld.signature.CryptoSuite;
-import com.apicatalog.ld.signature.SignatureSuite;
-import com.apicatalog.ld.signature.method.VerificationMethod;
-import com.apicatalog.ld.signature.proof.Proof;
+import com.apicatalog.vc.method.VerificationMethod;
+import com.apicatalog.vc.model.Proof;
+import com.apicatalog.vc.suite.SignatureSuite;
 
 import jakarta.json.JsonObject;
