+++ src/main/java/com/apicatalog/vc/method/VerificationMethod.java
@@ -1,4 +1,4 @@
-package com.apicatalog.ld.signature.method;
+package com.apicatalog.vc.method;
 
 import java.net.URI;
