+++ src/main/java/com/apicatalog/vc/model/Verifiable.java
@@ -3,8 +3,6 @@ package com.apicatalog.vc.model;
 import java.net.URI;
 import java.util.Collection;
 
-import com.apicatalog.ld.signature.proof.Proof;
-
 /**
  * Represents a common ancestor for verifiable data.
  *
