+++ src/main/java/com/apicatalog/vc/processor/Verifier.java
@@ -20,23 +20,23 @@ import com.apicatalog.ld.DocumentError.ErrorType;
 import com.apicatalog.ld.schema.LdTerm;
 import com.apicatalog.ld.signature.CryptoSuite;
 import com.apicatalog.ld.signature.LinkedDataSignature;
-import com.apicatalog.ld.signature.SignatureSuite;
-import com.apicatalog.ld.signature.SignatureSuiteProvider;
 import com.apicatalog.ld.signature.VerificationError;
 import com.apicatalog.ld.signature.VerificationError.Code;
 import com.apicatalog.ld.signature.key.VerificationKey;
-import com.apicatalog.ld.signature.method.DidUrlMethodResolver;
-import com.apicatalog.ld.signature.method.HttpMethodResolver;
-import com.apicatalog.ld.signature.method.MethodResolver;
-import com.apicatalog.ld.signature.method.VerificationMethod;
-import com.apicatalog.ld.signature.proof.EmbeddedProof;
-import com.apicatalog.ld.signature.proof.Proof;
 import com.apicatalog.vc.VcVocab;
 import com.apicatalog.vc.loader.StaticContextLoader;
+import com.apicatalog.vc.method.DidUrlMethodResolver;
+import com.apicatalog.vc.method.HttpMethodResolver;
+import com.apicatalog.vc.method.MethodResolver;
+import com.apicatalog.vc.method.VerificationMethod;
 import com.apicatalog.vc.model.Credential;
+import com.apicatalog.vc.model.EmbeddedProof;
+import com.apicatalog.vc.model.Proof;
 import com.apicatalog.vc.model.Verifiable;
 import com.apicatalog.vc.model.io.CredentialReader;
 import com.apicatalog.vc.model.io.PresentationReader;
+import com.apicatalog.vc.suite.SignatureSuite;
+import com.apicatalog.vc.suite.SignatureSuiteProvider;
 
 import jakarta.json.JsonArray;
 import jakarta.json.JsonObject;
